+++ app/src/main/java/com/mobdeve/s13/group1/budgetbuilder/Expense.kt
@@ -2,10 +2,10 @@ package com.mobdeve.s13.group1.budgetbuilder
 
 import java.util.*
 
-class Expense (
+class Expense(
     val date: Date,
     val type: String,
-    val amount: Int,
+    val amount: Float,
     val desc: String
     ) {
 }
