+++ app/src/main/java/com/mobdeve/s13/group1/budgetbuilder/ExpenseViewHolder.kt
@@ -41,7 +41,7 @@ class ExpenseViewHolder(view: View): RecyclerView.ViewHolder(view) {
         tvDesc.text = desc
     }
 
-    fun setAmount(amount: Int) {
+    fun setAmount(amount: Float) {
         tvExpenseAmount.text = "-" + amount.toString()
     }
